# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Produtos Natiops",
    "description": "API para buscar produtos e visualizar seus detalhes e estoque.",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "https://clientes-natiops.53rnlo.easypanel.host"
    }
  ],
  "paths": {
...
    "/holds": {
      "post": {
        "description": "Cria uma nova reserva de produto.",
        "operationId": "CriarReserva",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHoldDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reserva criada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Quantidade inválida ou estoque insuficiente."
          },
          "404": {
            "description": "Produto não encontrado."
          }
        },
        "deprecated": false
      }
    },
    "/holds/search": {
      "post": {
        "description": "Lista reservas ativas por email.",
        "operationId": "ListarReservasPorEmail",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindHoldsByEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reservas encontradas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindHoldsByEmailResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Email inválido ou ausente."
          }
        },
        "deprecated": false
      }
    },
    "/holds/{id}/cancel": {
      "patch": {
        "description": "Cancela uma reserva ativa.",
        "operationId": "CancelarReserva",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador da reserva (UUID).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "c1d2e3f4-5678-90ab-cdef-1234567890ab"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reserva cancelada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Reserva já foi finalizada ou inválida."
          },
          "404": {
            "description": "Reserva não encontrada."
          }
        },
        "deprecated": false
      }
    },
    "/holds/{id}/renew": {
      "patch": {
        "description": "Renova a data de expiração de uma reserva.",
        "operationId": "RenovarReserva",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador da reserva (UUID).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "c1d2e3f4-5678-90ab-cdef-1234567890ab"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenewHoldDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reserva renovada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Reserva não está ativa."
          },
          "404": {
            "description": "Reserva não encontrada."
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateHoldDto": {
        "type": "object",
        "properties": {
          "externalId": {
            "type": "string",
            "description": "Identificador externo do produto a ser reservado.",
            "example": "PROD-123"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email do cliente que está solicitando a reserva.",
            "example": "cliente@nativa.com.br"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "description": "Quantidade de itens que devem ser reservados.",
            "example": 2
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de expiração da reserva em ISO 8601.",
            "example": "2026-02-05T23:59:59Z"
          }
        },
        "required": [
          "externalId",
          "email",
          "quantity",
          "expiresAt"
        ]
      },
      "HoldResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "CANCELLED", "EXPIRED"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "productId",
          "email",
          "quantity",
          "expiresAt",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "FindHoldsByEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email usado para buscar reservas ativas.",
            "example": "cliente@nativa.com.br"
          }
        },
        "required": [
          "email"
        ]
      },
      "HoldWithProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "expiresAt": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "product": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "id",
          "email",
          "quantity",
          "expiresAt",
          "status",
          "createdAt",
          "updatedAt",
          "product"
        ]
      },
      "FindHoldsByEmailResponseDto": {
        "type": "object",
        "properties": {
          "holds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HoldWithProductResponseDto"
            }
          },
          "total": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "holds",
          "total"
        ]
      },
      "RenewHoldDto": {
        "type": "object",
        "properties": {
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Nova data de expiração da reserva em ISO 8601.",
            "example": "2026-02-10T23:59:59Z"
          }
        },
        "required": [
          "expiresAt"
        ]
      },
...
    }
  }
}